TM-4735 Modified the Account Export to reorder columns, formatted spreadsheet, formatted the review input form, added separate col index for spreadsheet and form, added Roles and Teams tabs.
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@6711 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ src/groovy/com/tdssrc/grails/WorkbookUtil.groovy
@@ -22,9 +22,10 @@ class WorkbookUtil {
 	}
 
 	public static addCell(sheet, columnIdx, rowIdx, value, type=null) {
-		def row = sheet.getRow(rowIdx)
+		//println "columnIdx=$columnIdx, rowIdx=$rowIdx, value=$value, rowIdx isa ${rowIdx.getClass().getName()}, sheet isa ${sheet.getClass().getName()}"
+		def row = sheet.getRow((int)rowIdx)
 		if (!row) {
-			row = sheet.createRow(rowIdx)
+			row = sheet.createRow((int)rowIdx)
 		}
 		def cell = row.getCell(columnIdx)
 		if (!cell) {
